Move scroll logic for sticky nav to higher-order component
Thanks to
https://medium.com/@dan_abramov/mixins-are-dead-long-live-higher-order-c
omponents-94a0d2f9e750#.vrh213sce for the example.
